examples/k8s_spark_plugin/requirements.txt
flytekitplugins-spark==v1.10.7b0
flytekitplugins-envd==v1.10.7b0
pandas
